requirements.txt
#list of  non robotframework
setuptools
lxml
selenium
#JayDeBeApi
pillow>=2.9.0
requests3
openpyxl
#csv file
unicodecsv
#google drive
gspread
oauth2client 
google-api-python-client
# support python
PyPDF2
pyyaml
#list of robotframework libraries
robotframework
robotframework-requests
robotframework-seleniumlibrary
robotframework-httplibrary
robotframework-databaselibrary



